feat(admin): relocate Platform Admin tooling into the Developer Toolkit
Move the org-override, product features / RBAC, and enterprise
onboarding controls out of the standalone OrgAdminSettings page (and
out of the org sidebar + user menu) into a new Platform Admin tab in
the dev-toolbar Developer Toolkit, so platform-admin tooling lives in
one place separate from normal org-admin settings.

- New PlatformAdminPanel renders the controls compactly for the
  toolbar; RBAC enable/disable uses an inline confirm instead of a
  modal (the toolbar collapses on outside clicks).
- Platform Admin tab is now visible to real platform admins (not just
  in dev); the dev-only impersonation toggle stays dev-gated.
- Remove the adminSettings route, the org-sidebar nav entry, the
  user-menu "Organization Override" item, and the OrgAdminSettings page.

Part 3 of AGE-2857.
