+++ client/dashboard/src/components/org-sidebar.tsx
@@ -13,7 +13,6 @@ import {
   SidebarMenu,
   SidebarMenuItem,
 } from "@/components/ui/sidebar";
-import { useIsPlatformAdmin } from "@/contexts/Auth";
 import { useTelemetry } from "@/contexts/Telemetry";
 import { Scope, useRBAC } from "@/hooks/useRBAC";
 import { AppRoute, useOrgRoutes } from "@/routes";
@@ -66,7 +65,6 @@ export function OrgSidebar({
   ...props
 }: React.ComponentProps<typeof Sidebar>): React.JSX.Element {
   const orgRoutes = useOrgRoutes();
-  const isAdmin = useIsPlatformAdmin();
   const { isRbacEnabled, isLoading: rbacLoading } = useRBAC();
   const telemetry = useTelemetry();
   const isDeviceAgentEnabled =
@@ -80,7 +78,6 @@ export function OrgSidebar({
     orgRoutes.domains,
     orgRoutes.logs,
     orgRoutes.webhooks,
-    orgRoutes.adminSettings,
   ].some((r) => r.active);
 
   const secureActive = [
@@ -112,7 +109,6 @@ export function OrgSidebar({
     orgRoutes.domains,
     orgRoutes.logs,
     orgRoutes.webhooks,
-    orgRoutes.adminSettings,
     orgRoutes.auditLogs,
     orgRoutes.deviceAgent,
     orgRoutes.access,
@@ -188,9 +184,6 @@ export function OrgSidebar({
                   item={orgRoutes.webhooks}
                   scope={["org:read", "org:admin"]}
                 />
-                {isAdmin && (
-                  <CollapsibleNavItem item={orgRoutes.adminSettings} />
-                )}
               </CollapsibleNavGroup>
 
               {/* Secure group */}
