+++ client/dashboard/src/components/sidebar-user-menu.test.tsx
@@ -3,7 +3,6 @@ import { afterEach, describe, expect, it, vi } from "vitest";
 
 vi.mock("@/contexts/Auth", () => ({
   useUser: () => ({ displayName: "Sagar", email: "s@x.dev", photoUrl: "" }),
-  useIsPlatformAdmin: () => false,
   useSession: () => ({ organizations: [{ id: "o1" }] }),
 }));
 vi.mock("@/contexts/Sdk", () => ({
@@ -17,7 +16,6 @@ vi.mock("@/routes", () => ({
   useRoutes: () => ({ settings: { goTo: vi.fn() } }),
   useOrgRoutes: () => ({
     billing: { goTo: vi.fn() },
-    adminSettings: { goTo: vi.fn() },
   }),
 }));
 vi.mock("react-router", () => ({
